+++ exporters/otlp/src/otlp_http_log_record_builder.cc
@@ -11,6 +11,7 @@
 #include "opentelemetry/exporters/otlp/otlp_http_log_record_builder.h"
 #include "opentelemetry/exporters/otlp/otlp_http_log_record_exporter_factory.h"
 #include "opentelemetry/exporters/otlp/otlp_http_log_record_exporter_options.h"
+#include "opentelemetry/sdk/configuration/http_tls_configuration.h"
 #include "opentelemetry/sdk/configuration/otlp_http_log_record_exporter_builder.h"
 #include "opentelemetry/sdk/configuration/otlp_http_log_record_exporter_configuration.h"
 #include "opentelemetry/sdk/configuration/registry.h"
@@ -34,6 +35,8 @@ std::unique_ptr<opentelemetry::sdk::logs::LogRecordExporter> OtlpHttpLogRecordBu
 {
   OtlpHttpLogRecordExporterOptions options(nullptr);
 
+  const auto *tls = model->tls.get();
+
   options.url                = model->endpoint;
   options.content_type       = OtlpBuilderUtils::ConvertOtlpHttpEncoding(model->encoding);
   options.json_bytes_mapping = JsonBytesMappingKind::kHexId;
@@ -44,10 +47,15 @@ std::unique_ptr<opentelemetry::sdk::logs::LogRecordExporter> OtlpHttpLogRecordBu
   options.http_headers =
       OtlpBuilderUtils::ConvertHeadersConfigurationModel(model->headers.get(), model->headers_list);
   options.ssl_insecure_skip_verify = false;
-  options.ssl_ca_cert_path         = model->certificate_file;
-  options.ssl_client_key_path      = model->client_key_file;
-  options.ssl_client_cert_path     = model->client_certificate_file;
-  options.compression              = model->compression;
+
+  if (tls != nullptr)
+  {
+    options.ssl_ca_cert_path     = tls->certificate_file;
+    options.ssl_client_key_path  = tls->client_key_file;
+    options.ssl_client_cert_path = tls->client_certificate_file;
+  }
+
+  options.compression = model->compression;
 
   return OtlpHttpLogRecordExporterFactory::Create(options);
 }